backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.1
python-dotenv==1.0.1
dj-database-url==2.2.0
django-cors-headers==4.4.0
gunicorn
